Refactored for legibility & clarity, added deltaTime to Render func to imply a generic ui & update call. Minor fixes to midi serialize/deserialize
